update
I have been tasked to create an utility app, of which is a vending machine program in Python. Within my program, it is just some simple concept but one of the main features would be the stock system function, which tracks and subtracts its stock by 1 per input, as well as the nested dictionaries, and also the code input function that take information from the nested dictionaries.